submodule update: pass options containing "[no-]"
"git submodule--helper update" understands these options, so append them
to `opts` and remove their intermediate variables.

As a result, we now pass every instance of the option and its "no-"
variant to "git submodule--helper update" instead of only passing the
last instance. As we noted in the previous commit, this is good because
this makes our options parsing consistent.

Signed-off-by: Glen Choo <chooglen@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>